refactor: move app.dart to app directory and update import paths
